artgallery/requirements.txt
Django>=5.0,<5.1
djangorestframework
djangorestframework-simplejwt
dj-rest-auth==6.0.0
django-allauth==0.63.0
django-filter
Pillow
psycopg2-binary
django-cors-headers
requests==2.32.3